requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
anthropic>=0.50.0
google-generativeai>=0.8.0
httpx>=0.27.0
pydantic==2.9.2
PyJWT>=2.8.0
pytest==8.3.3
